requirements.txt
biopython>=1.72
pandas>=1.4.1